requirements.txt
python-dotenv
peewee